# backend/backend-temp/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Turnario API",
    "description": "API para el sistema de gestión de citas y reservas Turnario",
    "version": "1.0.0",
    "contact": {
      "name": "Soporte Turnario",
      "email": "soporte@turnario.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Servidor de desarrollo"
    },
    {
      "url": "https://api.turnario.com/api/v1",
      "description": "Servidor de producción"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Registrar un nuevo usuario",
        "description": "Crea una nueva cuenta de usuario en el sistema",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegistration"
              },
              "example": {
                "fullName": "Juan Pérez",
                "email": "juan@example.com",
                "password": "password123",
                "userType": "client",
                "phone": "+1234567890",
                "dateOfBirth": "1990-01-01"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario registrado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Usuario ya existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Iniciar sesión",
        "description": "Autentica un usuario y devuelve tokens de acceso",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "email": "juan@example.com",
                "password": "password123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": ["Usuarios"],
        "summary": "Obtener lista de usuarios",
        "description": "Obtiene una lista paginada de usuarios (solo admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            },
            "description": "Número de página"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            },
            "description": "Elementos por página"
          },
          {
            "name": "userType",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["client", "professional", "admin"]
            },
            "description": "Filtrar por tipo de usuario"
          },
          {
            "name": "isActive",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "Filtrar por estado activo"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuarios obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponse"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Usuarios"],
        "summary": "Crear nuevo usuario",
        "description": "Crea un nuevo usuario (solo admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreation"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/services": {
      "get": {
        "tags": ["Servicios"],
        "summary": "Obtener lista de servicios",
        "description": "Obtiene una lista paginada de servicios",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            },
            "description": "Número de página"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            },
            "description": "Elementos por página"
          },
          {
            "name": "categoryId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "objectId"
            },
            "description": "Filtrar por categoría"
          },
          {
            "name": "professionalId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "objectId"
            },
            "description": "Filtrar por profesional"
          },
          {
            "name": "isActive",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "Filtrar por estado activo"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de servicios obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Servicios"],
        "summary": "Crear nuevo servicio",
        "description": "Crea un nuevo servicio (solo profesionales y admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceCreation"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Servicio creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "UserRegistration": {
        "type": "object",
        "required": ["fullName", "email", "password", "userType"],
        "properties": {
          "fullName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "description": "Nombre completo del usuario"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email del usuario"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "description": "Contraseña del usuario"
          },
          "userType": {
            "type": "string",
            "enum": ["client", "professional", "admin"],
            "description": "Tipo de usuario"
          },
          "phone": {
            "type": "string",
            "description": "Número de teléfono"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "description": "Fecha de nacimiento"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email del usuario"
          },
          "password": {
            "type": "string",
            "description": "Contraseña del usuario"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "user": {
                "$ref": "#/components/schemas/User"
              },
              "token": {
                "type": "string",
                "description": "JWT token"
              },
              "refreshToken": {
                "type": "string",
                "description": "Refresh token"
              }
            }
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "format": "objectId"
          },
          "fullName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "userType": {
            "type": "string",
            "enum": ["client", "professional", "admin"]
          },
          "phone": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "isEmailVerified": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "description": "Mensaje de error"
          },
          "error": {
            "type": "string",
            "description": "Tipo de error"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Detalles del error"
          }
        }
      },
...
      "UserListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "pages": {
            "type": "integer"
          }
        }
      },
      "Service": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "format": "objectId"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "duration": {
            "type": "integer",
            "description": "Duración en minutos"
          },
          "currency": {
            "type": "string",
            "enum": ["USD", "EUR", "COP"]
          },
          "professionalId": {
            "$ref": "#/components/schemas/User"
          },
          "categoryId": {
            "type": "string",
            "format": "objectId"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ServiceListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Service"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
...
    }
  },
...
}